2_TrainArchive/requirements.txt
beautifulsoup4==4.9.0
boto==2.49.0
boto3==1.13.11
botocore==1.16.11
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
Cython==0.29.14
docopt==0.6.2
docutils==0.15.2
dropbox==10.1.2
feedfinder2==0.0.4
feedparser==5.2.1
gcloud==0.17.0
gensim==3.6.0
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.5
jieba3k==0.35.1
jmespath==0.10.0
joblib==0.15.1
jws==0.1.3
logger==1.4
lxml==4.5.0
newspaper3k==0.2.8
nltk==3.5
num2words==0.5.10
numpy==1.18.4
oauth2client==3.0.0
pandas==1.0.3
Pillow==7.1.2
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.8.1
python-http-client==3.2.7
python-jwt==2.0.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.11.1
requests-file==1.5.1
requests-toolbelt==0.7.0
rsa==4.0
s3transfer==0.3.3
schedule==0.6.0
scikit-learn==0.23.0
scipy==1.4.1
sendgrid==6.3.1
six==1.14.0
sklearn==0.0
smart-open==2.0.0
soupsieve==2.0
threadpoolctl==2.0.0
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.46.0
urllib3==1.21.1
wincertstore==0.2
